+++ seminar05/hyeongmo/assignment5/model/Blog.js
@@ -3,8 +3,8 @@ const table = 'blog';
 
 module.exports = {
     create: async (json) => {
-        const fields = 'name, url';
-        const questions = `"${json.name}","${json.url}"`;
+        const fields = 'name, url, userIdx';
+        const questions = `"${json.name}","${json.url}", "${json.userIdx}"`;
         const result = await pool.queryParam_None(`INSERT INTO ${table}(${fields})VALUES(${questions})`)
         return result;
     },
@@ -17,11 +17,11 @@ module.exports = {
         return result;
     },
     update: async (json) => {
-        const result = await pool.queryParam_None(`UPDATE ${table} SET name = '${json.name}', url = '${json.url}' where blogIdx = '${json.blogIdx}'`)
+        const result = await pool.queryParam_None(`UPDATE ${table} SET name = '${json.name}', url = '${json.url}' where blogIdx = '${json.blogIdx}' and userIdx = '${json.userIdx}'`)
         return result;
     },
     remove: async (json) => {
-        const result = await pool.queryParam_None(`DELETE FROM ${table} WHERE blogIdx = '${json.blogIdx}'`)
+        const result = await pool.queryParam_None(`DELETE FROM ${table} WHERE blogIdx = '${json.blogIdx}' and userIdx = '${json.userIdx}'`)
         return result;
     }
 }
