+++ seminar05/hyeongmo/assignment5/model/Comment.js
@@ -4,8 +4,8 @@ const table = 'comment';
 
 module.exports = {
     create: async (json) => {
-        const fields = 'writer, content, articleIdx';
-        const questions = `"${json.writer}","${json.content}","${json.articleIdx}"`;
+        const fields = 'writer, content, articleIdx, userIdx';
+        const questions = `"${json.writer}","${json.content}","${json.articleIdx}", "${json.userIdx}"`;
         const result = await pool.queryParam_None(`INSERT INTO ${table}(${fields})VALUES(${questions})`)
         return result;
     },
@@ -18,11 +18,11 @@ module.exports = {
         return result;
     },
     update: async (json) => {
-        const result = await pool.queryParam_None(`UPDATE ${table} SET writer = '${json.writer}', content = '${json.content}' where commentIdx = '${json.commentIdx}' AND articleIdx = '${json.articleIdx}'`)
+        const result = await pool.queryParam_None(`UPDATE ${table} SET writer = '${json.writer}', content = '${json.content}' where commentIdx = '${json.commentIdx}' AND articleIdx = '${json.articleIdx}' AND userIdx = '${json.userIdx}'`)
         return result;
     },
     remove: async (json) => {
-        const result = await pool.queryParam_None(`DELETE FROM ${table} WHERE commentIdx='${json.commentIdx}' AND articleIdx='${json.articleIdx}'`)
+        const result = await pool.queryParam_None(`DELETE FROM ${table} WHERE commentIdx='${json.commentIdx}' AND articleIdx='${json.articleIdx}' AND userIdx='${json.userIdx}'`)
         return result;
     }
 }
